Exclude .git directory from deployment
Due to some repository having a big .git files, this change adds the exclusion of the .git directory to the list of patterns in the deployment process. This also ensures that sensitive version control information is not deployed along with the rest of the code, reducing the size of the deployment package.