Handle 'common' web file MIME types
This handles serving 'common' (js, css, html) file types within the static assets middleware.
The returned MIME type for these files is `text/plain` and based on this we attempt to determine the correct MIME type via their file extension.
This is by no means an exhaustive list but provides support for files which would be typical for a web application.